requirements.txt
torch
torchvision
tqdm
opencv-python
scikit-image
scikit-learn
numpy
Pillow
